// frontend/src/types/camera.ts
export interface Camera {
  id: number
  productCode: string
  name: string
  award?: string
  reviewScore?: number
  imageUrl?: string
  url?: string
  dprReviewArchiveUrl?: string
  executiveSummary?: string
  reviewGoodFor?: string
  reviewNotGoodFor?: string
  reviewConclusion?: string
  announced?: string
  bodyType?: string
  bodyMaterial?: string
  durability?: string
  dimensions?: string
  weightIncBatteries?: string
  environmentallySealed?: string
  processor?: string
  msrp?: string
  sensorType?: string
  sensorSize?: string
  effectivePixels?: string
  sensorPhotoDetectors?: string
  maxResolution?: string
  colorSpace?: string
  colorFilterArray?: string
  focalLengthEquiv?: string
  focalLengthMultiplier?: string
  opticalZoom?: string
  digitalZoom?: string
  maximumAperture?: string
  lensMount?: string
  iso?: string
  boostedIsoMaximum?: string
  boostedIsoMinimum?: string
  aperturePriority?: string
  shutterPriority?: string
  manualExposureMode?: string
  exposureCompensation?: string
  continuousDrive?: string
  maximumShutterSpeed?: string
  minimumShutterSpeed?: string
  manualFocus?: string
  autofocus?: string
  normalFocusRange?: string
  macroFocusRange?: string
  numberOfFocusPoints?: string
  builtInFlash?: string
  externalFlash?: string
  flashModes?: string
  flashRange?: string
  meteringModes?: string
  imageStabilization?: string
  imageRatioWh?: string
  jpegQualityLevels?: string
  uncompressedFormat?: string
  otherResolutions?: string
  screenSize?: string
  screenDots?: string
  screenType?: string
  touchScreen?: string
  articulatedLcd?: string
  viewfinderType?: string
  viewfinderCoverage?: string
  viewfinderMagnification?: string
  viewfinderResolution?: string
  videoFormat?: string
  resolutions?: string
  liveView?: string
  timelapseRecording?: string
  battery?: string
  batteryDescription?: string
  batteryLifeCipa?: string
  storageTypes?: string
  storageIncluded?: string
  usb?: string
  usbCharging?: string
  hdmi?: string
  headphonePort?: string
  microphonePort?: string
  wireless?: string
  wirelessNotes?: string
  gps?: string
  microphone?: string
  speaker?: string
  selfTimer?: string
  subjectSceneModes?: string
  remoteControl?: string
  orientationSensor?: string
  customWhiteBalance?: string
  whiteBalancePresets?: string
  wbBracketing?: string
}